bytes for memory limits need to be a psql bigint or the limit is 2.1GB

+++ atc/db/migration/migrations/1610989468_add_allocatable_resources_to_workers.up.sql
@@ -1,4 +1,4 @@
 BEGIN;
   ALTER TABLE workers ADD COLUMN allocatable_cpu integer;
-  ALTER TABLE workers ADD COLUMN allocatable_memory integer;
+  ALTER TABLE workers ADD COLUMN allocatable_memory bigint;
 COMMIT;
